+++ src/main/java/net/farmz/FarmClient.java
@@ -4,6 +4,7 @@ import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.farmz.init.RenderInit;
+import net.farmz.network.FarmClientPacket;
 
 @Environment(EnvType.CLIENT)
 public class FarmClient implements ClientModInitializer {
@@ -11,6 +12,7 @@ public class FarmClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         RenderInit.init();
+        FarmClientPacket.init();
     }
 
 }