+++ src/main/java/net/farmz/config/FarmConfig.java
@@ -2,7 +2,6 @@ package net.farmz.config;
 
 import me.shedaniel.autoconfig.ConfigData;
 import me.shedaniel.autoconfig.annotation.Config;
-import me.shedaniel.autoconfig.annotation.ConfigEntry;
 import me.shedaniel.cloth.clothconfig.shadowed.blue.endless.jankson.Comment;
 
 @Config(name = "farmz")
@@ -11,9 +10,6 @@ public class FarmConfig implements ConfigData {
 
     @Comment("Farmland water source range check")
     public int waterSourceRange = 0;
-    @ConfigEntry.BoundedDiscrete(min = 0, max = 100)
-    @Comment("Chance to plant a golden crop")
-    public int farmersHatGoldChance = 5;
 
     @Comment("Xp drop on breaking the plant")
     public int goldenPlantMinXp = 0;